Extract common functions to read/write PLY point clouds
