+++ src/util/ply.h
@@ -1,0 +1,50 @@
+// COLMAP - Structure-from-Motion and Multi-View Stereo.
+// Copyright (C) 2017  Johannes L. Schoenberger <jsch at inf.ethz.ch>
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+#ifndef COLMAP_SRC_UTIL_PLY_H_
+#define COLMAP_SRC_UTIL_PLY_H_
+
+#include <string>
+#include <vector>
+
+namespace colmap {
+
+struct PlyPoint {
+  float x = 0.0f;
+  float y = 0.0f;
+  float z = 0.0f;
+  float nx = 0.0f;
+  float ny = 0.0f;
+  float nz = 0.0f;
+  uint8_t r = 0;
+  uint8_t g = 0;
+  uint8_t b = 0;
+};
+
+// Read PLY point cloud from text or binary file.
+std::vector<PlyPoint> ReadPly(const std::string& path);
+
+// Write PLY point cloud to text or binary file.
+void WriteTextPly(const std::string& path, const std::vector<PlyPoint>& points,
+                  const bool write_normal = true, const bool write_rgb = true);
+void WriteBinaryPly(const std::string& path,
+                    const std::vector<PlyPoint>& points,
+                    const bool write_normal = true,
+                    const bool write_rgb = true);
+
+}  // namespace colmap
+
+#endif  // COLMAP_SRC_UTIL_PLY_H_
